fix(terminal): fix close-tab shortcut function
